Feat 관리자 매출관리 상세 매출 그래프 추가, null 예외처리

+++ src/main/java/kr/co/mgv/admin/sales/vo/TotalSales.java
@@ -15,5 +15,5 @@ public class TotalSales {
 	private Integer id;
 	private String name;
 	private String date;
-	private Integer totalSales;
+	private Long totalSales;
 }
